+++ rollup.config.lib.js
@@ -2,31 +2,18 @@ import typescript from 'rollup-plugin-typescript2'
 import { dts } from "rollup-plugin-dts"
 
 // Cordova Library Configuration
+const expectedCordovaModules = ["cordova-powerauth-mobile-sdk", "cordova", "cordova-powerauth-networking", "cordova-digital-onboarding", "iproov-cordova-plugin", "blinkid-cordova-plugin"]
+
+// Cordova Library Configuration
 const libCordovaDir = 'packages/lib-cordova'
 const libCordovaInput = `${libCordovaDir}/src/index.ts`
 const libCordovaOutput = `${libCordovaDir}/lib/index.js`
 const libCordovaOutputDts = `${libCordovaDir}/lib/index.d.ts`
-const expectedCordovaModules = ["cordova-powerauth-mobile-sdk", "cordova", "cordova-powerauth-networking", "cordova-digital-onboarding", "iproov-cordova-plugin", "blinkid-cordova-plugin"]
-
-// Cordova App Configuration
-const appCordovaDir = 'testapp-cordova'
 
 // React Native Library Configuration
 const libRNDir = 'packages/lib-react-native'
 const libRNInput = `${libRNDir}/src/index.ts`
-const libRNOutput = `${libRNDir}/lib` 
-
-// We dont want to import modules that will be supplied by Cordova environment
-// Cordova plugins are injected at runtime, so we need to strip these imports from the final bundle to not cause errors.
-const stripCordovaImportsPlugin =  {
-  name: "remove-cordova-modules",
-  transform(code, id) {
-    return {
-      code: code.replace(new RegExp(`^import.*(?:${expectedCordovaModules.map(m => `"${m}"`).join("|")}).*$`, "gm"), ""),
-      map: null,
-    };
-  },
-}
+const libRNOutput = `${libRNDir}/lib`
 
 // Plugin to strip copyright comments from .d.ts files
 const stripCopyrightPlugin = {
@@ -94,21 +81,6 @@ export default [
       dts({ compilerOptions: { stripInternal: true } }),
       stripCopyrightPlugin
     ],
-  },
-  // Cordova App
-  {
-    input: `${appCordovaDir}/src/index.ts`,
-    output: {
-      file: `${appCordovaDir}/www/js/index.js`,
-      format: 'cjs',
-      sourcemap: true
-    },
-    plugins: [
-      typescript({
-        tsconfig: `${appCordovaDir}/tsconfig.json`
-      }),
-      stripCordovaImportsPlugin
-    ]
   }
 
   /**************
